Dark theme redesign, interactive plots, and publication export
- Add modern dark theme with accent colors via QApplication stylesheet
- Rearrange video previews to top/bottom layout for larger display
- Add matplotlib navigation toolbars for pan/zoom on plots
- D(t) plot uses sliding time window with auto-scroll
- Export saves publication-quality PNGs (white bg, 300 DPI)
- Fix memory error in box_count by using uint8 instead of int64
- Add README with usage instructions and method explanations

+++ src/core.py
@@ -55,8 +55,8 @@ class FractalAnalyzer:
         if binary_image is None or np.sum(binary_image) == 0:
             return 0.0, 0.0, [], [], False
 
-        # Ensure binary 0/1
-        pixels = (binary_image > 0).astype(int)
+        # Ensure binary 0/1 (use uint8 to minimize memory)
+        pixels = (binary_image > 0).astype(np.uint8)
         
         # Minimal dimension
         H, W = pixels.shape
